Three actions
This code has 3 steps.
Firstly, it checks if everybody is home, code starts. (I write a complement of a ping value for iPhone because if iPhone is connected a Wifi but the phone is locked, the value of ping becomes false.)
Secondly, if the motion sensor catches a motion, it calls a function.
Lastly, the function calculates the distance of the thing which is moving.